feat: move the src files to src directory
