+++ src/nanogpt_seq2seq/bpe_shakespeare_vocab300.json
