+++ src/nanogpt_seq2seq/generate.py
