+++ src/nanogpt_seq2seq/shakespeare.txt
