+++ src/nanogpt_seq2seq/simplified_nanogpt.py
